Added inputlink property in Carousel, Cards, Link and Teaserprimary components
