+++ fragments/cardthreeimages/template.html
@@ -1,52 +1,45 @@
 <div>
-
-    <div class="first" style="  width: 100%;
+  <div class="first" style="  width: 100%;
   height: 700px;">
-        <div class="second" style=" float: left;
+    <div class="second" style=" float: left;
   width: 50%;
   height:700px;
   margin-left:2px;">
-            <div>
-                <h2>Hello india</h2>
-                <p>Hello all good morning</p>
-            </div>
-            <img src="https://demo.vuestorefront.io/assets/ban1.jpg" style=" height: 590px;
+      <div>
+        <h2>Lorem ipsum</h2>
+        <p>lorem ipsum.</p>
+      </div>
+        <img src="ban1.jpg" style=" height: 590px;
   width: 100%;
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;">
-        </div>
-        <div class="second-two" style=" float: right;
+    </div>
+    <div class="second-two" style=" float: right;
   width: 47%;
   height:700px;">
-            <div>
-                <div>
-                    <h2>hello all</h2>
-                    <p>hello all</p>
-                </div>
-                <img src="https://demo.vuestorefront.io/assets/ban1.jpg" style=" height: 250px;
-  width: 100%;
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;"/>
-            </div>
-            <div>
-                <div>
-                    <h2>hello all</h2>
-                    <p>hello all</p>
-                </div>
-                <img src="https://demo.vuestorefront.io/assets/ban1.jpg" style=" height: 250px;
-  width: 100%;
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;"/>
-
-
-            </div>
-
-
-
+      <div>
+        <div>
+          <h2>Lorem ipsum</h2>
+          <p>lorem ipsum.</p>
         </div>
-
+          <img src="ban1.jpg" style=" height: 250px;
+  width: 100%;
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;"/>
+      </div>
+      <div>
+        <div>
+          <h2>Lorem ipsum</h2>
+          <p>lorem ipsum.</p>
+        </div>
+          <img src="ban1.jpg" style=" height: 250px;
+  width: 100%;
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;"/>
+      </div>
     </div>
+  </div>
 </div>
