+++ fragments/link/hatch.js
@@ -1,0 +1,7 @@
+module.exports = {
+    convert: function($, f) {
+        f.bindPath($)
+        f.mapRichField($.find('p'), "model.title")
+        f.mapRichField($.find('a').first(), 'href', "model.link")
+    }
+}
