+++ fragments/link/model.json
@@ -1,0 +1,23 @@
+{
+  "definitions": {
+    "Links": {
+      "type": "object",
+      "x-type": "component",
+      "properties": {
+        "title": {
+          "type": "string",
+          "x-source": "inject",
+          "x-form-label": "Image Title",
+          "x-form-type": "text"
+        },
+        "link": {
+          "type": "string",
+          "x-source": "inject",
+          "x-form-label": "link",
+          "x-form-type": "inputlink",
+          "x-form-browserRoot": "/context/sites"
+        }
+      }
+    }
+  }
+}
