+++ fragments/link/sample.json
@@ -1,0 +1,7 @@
+{
+  "title": "Link",
+  "group": "",
+  "model": {
+        "title": "Demo link"
+  }
+}
