+++ fragments/link/template.html
@@ -1,5 +1,3 @@
 <div>
-    <div>
     <a href=""><p> example </p></a>
-    </div>
 </div>
