--- fragments/links/sample.json
@@ -1,13 +1,0 @@
-{
-  "title": "Links",
-  "group": "",
-  "model": {
-    "links": [
-      {
-        "title": "Demo link 1",
-        "internallink": "",
-        "externalurl": "https://www.google.com"
-      }
-    ]
-  }
-}
