+++ ui.apps/src/main/content/jcr_root/apps/themecleanflex/components/link/.content.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0" jcr:primaryType="per:Component" jcr:title="Link" group="">
+  <jcr:content jcr:primaryType="nt:unstructured" title="Demo link"/>
+</jcr:root>
