+++ ui.apps/src/main/content/jcr_root/apps/themecleanflex/components/link/template.vue
@@ -1,0 +1,12 @@
+<template>
+  <div v-bind:data-per-path="model.path">
+    <a v-bind:href="model.link"><p v-html="model.title"></p></a>
+  </div>
+</template>
+
+<script>
+    export default {
+        props: ['model']
+    }
+</script>
+
